feat: Add id field to LobbyCreate model
The id field is added to the LobbyCreate model in the game_model.py file. This field represents the unique identifier for the lobby. It is set to a default empty string value.

Deleted CreateLobbyView.vue component

The CreateLobbyView.vue component is deleted from the frontend/src/views directory. This component was no longer needed and has been removed.

Update index.html with Montserrat font

The index.html file in the frontend directory is updated to include the Montserrat font from Google Fonts. This font will be used in the application.

Update router configuration in index.js

The router configuration in the index.js file in the frontend/src/router directory is updated. The HomeView component is replaced with the HomeComponent component. Additionally, the LegalNoticeComponent and WaitingLobbyComponent are added as new routes.

Deleted CreateLobbyComponent.vue component

The CreateLobbyComponent.vue component is deleted from the frontend/src/components directory. This component was no longer needed and has been removed.